file split MessageBox ScrollBar
