+++ chat-app/package.json
@@ -12,6 +12,7 @@
     "firebase": "^7.19.0",
     "register-service-worker": "^1.7.1",
     "vue": "^3.0.0-0",
+    "vue-infinite-loading": "^2.4.5",
     "vue-router": "^4.0.0-0",
     "vuex": "^4.0.0-0"
   },
